Move IterVisitor derived classes from fusion.h to iter_visitor.h
