+++ torch/csrc/jit/codegen/cuda/iter_visitor.h
@@ -4,6 +4,11 @@
 
 #include <torch/csrc/jit/codegen/cuda/dispatch.h>
 
+#include <torch/csrc/jit/codegen/cuda/fusion.h>
+#include <torch/csrc/jit/codegen/cuda/ir_all_nodes.h>
+#include <torch/csrc/jit/codegen/cuda/ir_iostream.h>
+#include <torch/csrc/jit/codegen/cuda/type.h>
+
 #include <deque>
 #include <unordered_set>
 #include <vector>
@@ -11,14 +16,6 @@
 namespace torch {
 namespace jit {
 namespace fuser {
-
-class Statement;
-class Val;
-class Expr;
-
-class Fusion;
-
-enum class ValType;
 
 /*
  * IterVisitor starts from leaf nodes, fusion outputs, or the provided values.
@@ -49,9 +46,31 @@ class TORCH_CUDA_API IterVisitor : public OptOutDispatch {
   // These functions will start at outputs and propagate up through the DAG
   // to inputs based on depth first traversal. Next could be called on a node
   // multiple times.
-  virtual std::vector<Statement*> next(Statement* stmt);
-  virtual std::vector<Statement*> next(Expr* expr);
-  virtual std::vector<Statement*> next(Val* v);
+  virtual inline std::vector<Statement*> next(Statement* stmt) {
+    if (stmt->isVal()) {
+      return next(stmt->as<Val>());
+    } else if (stmt->isExpr()) {
+      return next(stmt->as<Expr>());
+    } else {
+      TORCH_INTERNAL_ASSERT(
+          false, "IterVisitor could not detect type in next_dispatch.");
+    }
+  }
+
+  virtual inline std::vector<Statement*> next(Val* v) {
+    FusionGuard::getCurFusion()->assertInFusion(v, "Cannot traverse val, ");
+    if (FusionGuard::getCurFusion()->origin(v) != nullptr) {
+      return {FusionGuard::getCurFusion()->origin(v)};
+    }
+    return {};
+  }
+
+  virtual inline std::vector<Statement*> next(Expr* expr) {
+    FusionGuard::getCurFusion()->assertInFusion(expr, "Cannot traverse expr, ");
+    std::vector<Statement*> next_stmts{expr->inputs().begin(),
+                                       expr->inputs().end()};
+    return next_stmts;
+  }
 
   // This handle functions is called on every Statement* in topological order,
   // starting from outputs to inputs.
@@ -212,6 +231,32 @@ class TORCH_CUDA_API DependencyCheck {
   static std::unordered_set<Val*> getAllValsBetween(
       const std::unordered_set<Val*>& dependencies,
       const std::vector<Val*>& of);
+};
+
+// Expr sort will take a fusion and return a topologically sorted list of
+// expressions.
+class ExprSort : public IterVisitor {
+ private:
+  std::vector<Expr*> exprs;
+
+  void handle(Expr* expr) override;
+
+ public:
+  static std::vector<Expr*> getExprs(Fusion* fusion, bool from_outputs_only);
+
+  static std::vector<Expr*> getExprs(
+      Fusion* fusion,
+      const std::vector<Val*>& from);
+};
+
+class InputsOf : public IterVisitor {
+ private:
+  std::unordered_set<Val*> inputs;
+
+  void handle(Val* v) final;
+
+ public:
+  static std::unordered_set<Val*> output(Fusion* fusion, Val* output_);
 };
 
 } // namespace fuser
