+++ torch/csrc/jit/codegen/cuda/type.h
@@ -13,6 +13,14 @@ namespace torch {
 namespace jit {
 namespace fuser {
 
+// https://stackoverflow.com/questions/18837857/cant-use-enum-class-as-unordered-map-key
+struct TypeHash {
+  template <typename T>
+  std::size_t operator()(T t) const {
+    return static_cast<std::size_t>(t);
+  }
+};
+
 // Order of strength
 enum class ValType {
   TensorDomain,
